Move 'contains' to private folder
