+++ addon/-private/execution_context/acceptance.js
@@ -1,13 +1,23 @@
+import {
+  findElementWithAssert
+} from '../../helpers';
+
 export default function AcceptanceExecutionContext(pageObjectNode) {
   this.pageObjectNode = pageObjectNode;
 }
 
 AcceptanceExecutionContext.prototype = {
   run(cb) {
+    return cb(this);
+  },
+
+  runAsync(cb) {
     /* global wait */
     wait().then(() => {
       cb(this);
     });
+
+    return this.pageObjectNode;
   },
 
   click(selector, container) {
@@ -27,5 +37,9 @@ AcceptanceExecutionContext.prototype = {
   triggerEvent(selector, container, eventName, eventOptions) {
     /* global triggerEvent */
     triggerEvent(selector, container, eventName, eventOptions);
+  },
+
+  findWithAssert(selector, options) {
+    return findElementWithAssert(this.pageObjectNode, selector, options);
   }
 };
